synth-factory/requirements.txt
# Synthetic Data Factory - Dependencies
# Install: pip install -r requirements.txt

# Core data generation
sdv>=1.12.0
faker>=28.0.0

# Data handling
pandas>=2.2.0
numpy>=1.26.0
pyarrow>=15.0.0

# LLM integration (Ollama)
requests>=2.31.0
httpx>=0.27.0

# Validation & quality
evidently>=0.4.0

# Configuration
pyyaml>=6.0.1
pydantic>=2.6.0

# Packaging & reporting
jinja2>=3.1.3

# German-specific utilities
schwifty>=2024.1.1  # IBAN validation

# Testing
pytest>=8.0.0
